⚠️ conflicts — Upstream merge → 2026-06-20 (#16)
* vendor: soh develop 74e1d4c20

* chore(upstream): bump pins + merge log 2026-06-20

* resolved upstream merge conflicts (soh Network)

* fix(libultraship): back-port Kenix3 #1126 warning fixes for soh /WX

soh@develop builds with warnings-as-errors (#6733) and trips on two LUS headers
we vendor (lus_gbi.h GIMMCMD redefinition, BitConverter.h size_t->int conversion).
soh tracks LUS port-maintenance (has #1126); we vendor LUS main (doesn't yet), so
back-port just #1126 additively. Did NOT switch to port-maintenance (carries #1103,
intentionally skipped). Documented in docs/merges/2026-06-20.md.

* fix(combo): correct AddRectFilled args tripping soh /WX (C4244)

Two ComboMenuDrawContent.h calls passed 'true' as the float rounding arg and
style.WindowRounding (a float) into the ImDrawFlags slot — a float->ImDrawFlags
conversion that soh's warnings-as-errors now rejects. Drop the stray 'true' so
WindowRounding lands in the rounding arg (the intended behavior) with default
flags. Combo-owned code; no guard needed.

* build(combo): silence DLL-boundary warnings on shared soh/2ship under /WX

ComboShip builds soh/2ship as SHARED DLLs (upstream builds exes), so MSVC emits
DLL-boundary warnings upstream never sees, and soh's new /WX (#6733) makes them
fatal — starting with C4273 'inconsistent dll linkage' on the floorf/ceilf
forward-decls in OTRGlobals.cpp. Disable that class (C4273/C4251/C4275) for the
shared game targets from the combo layer (no vendored CMake edits); /WX still
catches genuine warnings. Systemic fix so this class stops recurring per merge.

+++ soh/soh/Network/Network.cpp
@@ -82,7 +82,7 @@ void Network::SendDataToRemote(const char* payload) {
     }
 #else
     SPDLOG_DEBUG("[Network] Sending data: {}", payload);
-    SDLNet_TCP_Send(networkSocket, payload, strlen(payload) + 1);
+    SDLNet_TCP_Send(networkSocket, payload, static_cast<int>(strlen(payload) + 1));
 #endif
 }
